feat: add useCurrentUser hook and refactor profile
