+++ src/services/index.ts
@@ -1,6 +1,7 @@
 export { AuthService } from './auth/AuthService';
 export { TokenService } from './auth/TokenService';
 export { ProfileService } from './profile/ProfileService';
+export { useCurrentUser } from './profile/useCurrentUser';
 export { HttpClient } from './core/HttpClient';
 export { useAuth } from './auth/useAuth';
